Move slate element types to type.ts file
